Chore(messaging): refactored routes (#546)
* chore(messaging): refactored routes

* updated sdks

* Fixed sdks
