+++ apps/messaging-api/routes/index.ts
@@ -5,6 +5,7 @@ import emails from "./providers/emails";
 import sms from "./providers/sms";
 import templates from "./templates";
 import users from "./users";
+import usersImports from "./users/imports";
 
 export default async function routes(app: FastifyInstance) {
   app.register(messages, { prefix: "/messages" });
@@ -12,6 +13,7 @@ export default async function routes(app: FastifyInstance) {
   app.register(sms, { prefix: "/providers/sms" });
   app.register(templates, { prefix: "/templates" });
   app.register(users, { prefix: "/users" });
+  app.register(usersImports, { prefix: "/users/imports" });
 }
 
 export interface SMSService {
